refactor: move util to own directory
